Fix initial memcpy in strcat operator
We were doing a memcpy of the total allocated size minus 1, presumably because
we knew that the string would be null terminated and we didn't want to copy the
null.  However, strcat requires both of its arguments to be null terminated, so
if we don't copy over the null terminator from str1, then we are only guaranteed
that our first argument to strcat will be null terminated in the case where we
re-Calloc'ed the buffer.  Otherwise, we might run into danger.

With this change, the attached code seems to run without errors.