Fix handling when opcode shadows variable
Since the fix for differentiating between opcodes with a parenthesized
first argument and function starts, there was a transformation in
`convert_statement_to_opcall()` that was no longer getting called in
its intended situation.  However, it was getting called and performing
an invalid transformation on other trees.  It has been removed.

Also, in cases where we have a line of the form `ident1 ident2`, we
were not checking whether ident2 was a variable and respecting that,
but instead just checking whether it was an opcode.  There is now
a check for looking through the variable pools.

Tests were added to the ambigous opcall csdtest to cover these cases.