Fix for ambiguous array expressions
When a variable was shadowing a UDO that returned an array, we would
always try to lookup the name of the UDO in the variable pools first,
even though we knew it was a T_FUNCTION node.  Now, when we see a
T_FUNCTION, we check first to see if it's a UDO that returns an array.